Add tabindex attribute to searchbar
Summary:
Explain the **motivation** for making this change. What existing problem does the pull request solve?

This is small improvement to docs webpage.
I'm tired of clicking on searchbar with mouse/touchpad, so I've added a `tabindex` attribute to it, so it can be now focused with `Tab` key.

I think this greatly improves user experience on documentation searching.

**Test plan (required)**

??? Just click on `Tab` and it will focus. Should I test it really?

**Code formatting**

Attributes are sorted alphabetically, I've inserted new attribute in this order, after `id` and before `type` attributes.
Closes https://github.com/facebook/react-native/pull/8319

Differential Revision: D3470858

fbshipit-source-id: 01240bdf1432d9873324ddee38be256dfab93df2